feat:安装模版到项目/模版实例化时，自动安装模版中需安装的容器镜像 #3756
